+++ src/phoenix/trace/__init__.py
@@ -1,4 +1,4 @@
+from .span_evaluations import SpanEvaluations
 from .trace_dataset import TraceDataset
-from .trace_evaluations import TraceEvaluations
 
-__all__ = ["TraceDataset", "TraceEvaluations"]
+__all__ = ["TraceDataset", "SpanEvaluations"]
