--- src/phoenix/trace/trace_evaluations.py
@@ -1,111 +1,0 @@
-from collections import OrderedDict
-from typing import Union
-
-import pandas as pd
-
-EVALUATIONS_INDEX_NAME = "context.span_id"
-
-EVAL_NAME_COLUMN_PREFIX = "eval."
-
-
-class TraceEvaluations:
-    """
-    TraceEvaluations is a set of evaluation annotations for a set of spans.
-    TraceEvaluations encompasses the evaluation annotations for a single evaluation task
-    such as toxicity or hallucinations.
-
-    TraceEvaluations can be appended to TraceDatasets so that the spans and
-    evaluations can be joined and analyzed together.
-
-    Parameters
-    __________
-    eval_name: str
-        the name of the evaluation, e.x. 'toxicity'
-    dataframe: pandas.DataFrame
-        the pandas dataframe containing the evaluation annotations Each row
-        represents the evaluations on a span.
-
-    Example
-    _______
-
-    DataFrame of evaluations for toxicity may look like:
-
-    | span_id | score              | label              | explanation        |
-    |---------|--------------------|--------------------|--------------------|
-    | span_1  | 1                  | toxic              | bad language       |
-    | span_2  | 0                  | non-toxic          | violence           |
-    | span_3  | 1                  | toxic              | discrimination     |
-    """
-
-    dataframe: pd.DataFrame
-
-    eval_name: str  # The name for the evaluation, e.x. 'toxicity'
-
-    def __init__(self, eval_name: str, dataframe: pd.DataFrame):
-        self.eval_name = eval_name
-        # validate that the dataframe is indexed by context.span_id
-        if dataframe.index.name != EVALUATIONS_INDEX_NAME:
-            raise ValueError(
-                f"The dataframe index must be '{EVALUATIONS_INDEX_NAME}' but was "
-                f"'{dataframe.index.name}'"
-            )
-        self.dataframe = dataframe
-
-    def get_dataframe(self, prefix_columns_with_name: bool = True) -> pd.DataFrame:
-        """
-        Returns a copy of the dataframe with the evaluation annotations
-
-        Parameters
-        __________
-        prefix_columns_with_name: bool
-            if True, the columns will be prefixed with the eval_name, e.x. 'eval.toxicity.value'
-        """
-        dataframe = self.dataframe.copy()
-        if prefix_columns_with_name:
-            # Prefix all columns except the span_id column
-            columns_to_prefix = [
-                column for column in dataframe.columns if column != EVAL_NAME_COLUMN_PREFIX
-            ]
-            dataframe.rename(
-                columns={
-                    column: f"{EVAL_NAME_COLUMN_PREFIX}{self.eval_name}.{column}"
-                    for column in columns_to_prefix
-                },
-                inplace=True,
-            )
-        return dataframe
-
-
-def binary_classifications_to_trace_evaluations(
-    eval_name: str,
-    classifications_df: pd.DataFrame,
-    rails_map: "OrderedDict[bool, str]",
-) -> TraceEvaluations:
-    """
-    Takes a dataframe of binary classifications and converts it to a TraceEvaluations
-
-    Parameters
-    __________
-    eval_name: str
-        the name of the evaluation, e.x. 'toxicity'
-    classifications_df: pd.DataFrame
-        the dataframe of binary classifications, typically the output of llm evals
-    spans_df: pd.DataFrame
-        the dataframe of spans over which the classifications were made
-    rails_map: OrderedDict[bool, str]
-        a map of the binary classifications to the labels. E.x. {True: "toxic", False: "non-toxic"}
-    """
-    evaluations_df = classifications_df.copy()
-
-    # Convert the rails map to a map of labels to binary values
-    classification_map = {label: binary for binary, label in rails_map.items()}
-
-    # Use the evaluations to convert the binary classifications to a 1 or 0 value
-    def map_classification_to_value(classification: str) -> Union[int, None]:
-        if classification_map.get(classification) is None:
-            return None
-        return 1 if classification_map[classification] else 0
-
-    evaluations_df["value"] = evaluations_df["label"].apply(map_classification_to_value)
-
-    return TraceEvaluations(eval_name, evaluations_df)
